Reduce likelihood of large oscillations in predictions
A major change in the signal can cause issues for the forecasting,
as it can easily overcompensate while trying to adjust to the
new signal pattern.

These fixes mostly try to "soften" the values that we save in
our history when there is a big difference between an observed
value and what we had forecast.

Specifically, we try to err on the side of updating based on the
value (forecast or observed) that was closer to the previous
observation, rather than always trusting the forecast. This means
that when our forecasting starts to overcompensate, we do not
keep compounding the forecasting error and instead start to bring
things back into line with the observed values.

Other changes:
 * when shifting / adjusting the forecasted value to insert as
   saved history, use 20% of the difference between forecasted
   and observed OR 20% of the forecast, whichever is smaller.
   This fixes an issue where a large brief spike would cause a
   dramatic shift in future predictions that took a long time to
   recover from.
 * account for potentially negative forecasted values -- do NOT
   try to insert them into history, or attempt the adjustment
   calculations (as a negative value here will push the adjustment
   in the wrong direction anyway).