Only re-generate a model if we're processing newer data
Any historical / backfilling processes should not replace any
model derived from live data, as this may produce a worse model
for the live processing.

Especially, if the historical processing starts close to where
we first started collecting data, as any model will be then be
based on history with a lot of missing data points!