Remove code that writes log and result files when building models
We do not use these logs and having to make sure that the
directories exist to write these files into is a pain.

Any important logs are now written to syslog, so we can still
see them (and for debugging, you can always change the log level
if you need the other logs messages as well).